+++ TDEngine2/plugins/TDE2LevelStreamingUtils/source/CSceneChunksLoadingSystem.cpp
@@ -1,6 +1,7 @@
 #include "../include/CSceneChunksLoadingSystem.h"
 #include "../include/CSceneLoadingTriggerComponent.h"
 #include <scene/ISceneManager.h>
+#include <scene/IScene.h>
 #include <ecs/IWorld.h>
 #include <ecs/CTransform.h>
 #include <ecs/components/CBoundsComponent.h>
@@ -60,8 +61,39 @@ namespace TDEngine2
 	}
 
 	void CSceneChunksLoadingSystem::Update(IWorld* pWorld, F32 dt)
-	{
+	{	
+		auto& sceneLoadingTriggers = std::get<std::vector<CSceneLoadingTriggerComponent*>>(mContext.mComponentsSlice);
+		auto& bounds               = std::get<std::vector<CBoundsComponent*>>(mContext.mComponentsSlice);
+		auto& transforms           = std::get<std::vector<CTransform*>>(mContext.mComponentsSlice);
 
+		/// \note First step is updating of all bounds
+		for (USIZE i = 0; i < mContext.mComponentsCount; ++i)
+		{
+			auto&& pSceneLoadingTrigger = sceneLoadingTriggers[i];
+
+			const TVector3& sizes = pSceneLoadingTrigger->GetVolumeSizes();
+
+			const TAABB boundingVolume(transforms[i]->GetPosition() + pSceneLoadingTrigger->GetVolumeOffset(), sizes.x, sizes.y, sizes.z);
+			bounds[i]->SetBounds(boundingVolume);
+
+			if (ContainsPoint(boundingVolume, mCurrActiveCameraTransform->GetPosition()))
+			{
+				auto&& scenePath = pSceneLoadingTrigger->GetScenePath();
+
+				if (!scenePath.empty())
+				{
+					mpSceneManager->LoadSceneAsync(pSceneLoadingTrigger->GetScenePath(), {});	/// \note Load scene's chunk
+				}
+			}
+			else
+			{				
+				const TSceneId sceneHandle = mpSceneManager->GetSceneId(pSceneLoadingTrigger->GetScenePath());
+				if (TSceneId::Invalid != sceneHandle)
+				{
+					mpSceneManager->UnloadScene(sceneHandle); /// \note Unload it
+				}
+			}
+		}
 	}